# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API (apps/api)",
    "version": "1.0.0",
    "description": "OpenAPI spec for apps/api"
  },
  "servers": [
    {
      "url": "https://osc2zjnl43.execute-api.us-east-1.amazonaws.com"
    }
  ],
  "paths": {
...
    "/master": {
      "get": {
        "summary": "Retrieve a list of masters",
        "tags": [
          "Master"
        ],
        "responses": {
          "200": {
            "description": "A list of masters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Master"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new master (ADMIN only)",
        "tags": [
          "Master"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created master",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Master"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/master/{id}": {
      "put": {
        "summary": "Update an existing master (ADMIN only)",
        "tags": [
          "Master"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the master to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated master",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Master"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid id"
          },
          "404": {
            "description": "Master not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a master (ADMIN only)",
        "tags": [
          "Master"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Numeric ID of the master to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "No content (deleted)"
          },
          "400": {
            "description": "Invalid id"
          },
          "404": {
            "description": "Master not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Master": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Senior nail technician"
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "Presigned URL or null"
          }
        }
      },
      "MasterCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "S3 key returned from upload endpoint"
          }
        }
      },
      "MasterUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "S3 key returned from upload endpoint"
          }
        }
      },
...
    }
  },
...
}